// src/clients/weatherClient.ts
export class ForecastTimeStep implements IForecastTimeStep {
    /** Forecast for a specific time */
    data!: Data;
    /** The time these forecast values are valid for. Timestamp in format YYYY-MM-DDThh:mm:ssZ (ISO 8601) */
    time!: string;

    constructor(data?: IForecastTimeStep) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
        if (!data) {
            this.data = new Data();
        }
    }

    init(_data?: any) {
        if (_data) {
            this.data = _data["data"] ? Data.fromJS(_data["data"]) : new Data();
            this.time = _data["time"];
        }
    }

    static fromJS(data: any): ForecastTimeStep {
        data = typeof data === 'object' ? data : {};
        let result = new ForecastTimeStep();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["data"] = this.data ? this.data.toJSON() : <any>undefined;
        data["time"] = this.time;
        return data;
    }
}

export interface IForecastTimeStep {
    /** Forecast for a specific time */
    data: Data;
    /** The time these forecast values are valid for. Timestamp in format YYYY-MM-DDThh:mm:ssZ (ISO 8601) */
    time: string;
}

export class ForecastUnits implements IForecastUnits {
    air_pressure_at_sea_level?: string | undefined;
    air_temperature?: string | undefined;
    air_temperature_max?: string | undefined;
    air_temperature_min?: string | undefined;
    cloud_area_fraction?: string | undefined;
    cloud_area_fraction_high?: string | undefined;
    cloud_area_fraction_low?: string | undefined;
    cloud_area_fraction_medium?: string | undefined;
    dew_point_temperature?: string | undefined;
    fog_area_fraction?: string | undefined;
    precipitation_amount?: string | undefined;
    precipitation_amount_max?: string | undefined;
    precipitation_amount_min?: string | undefined;
    probability_of_precipitation?: string | undefined;
    probability_of_thunder?: string | undefined;
    relative_humidity?: string | undefined;
    ultraviolet_index_clear_sky_max?: string | undefined;
    wind_from_direction?: string | undefined;
    wind_speed?: string | undefined;
    wind_speed_of_gust?: string | undefined;

    constructor(data?: IForecastUnits) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.air_pressure_at_sea_level = _data["air_pressure_at_sea_level"];
            this.air_temperature = _data["air_temperature"];
            this.air_temperature_max = _data["air_temperature_max"];
            this.air_temperature_min = _data["air_temperature_min"];
            this.cloud_area_fraction = _data["cloud_area_fraction"];
            this.cloud_area_fraction_high = _data["cloud_area_fraction_high"];
            this.cloud_area_fraction_low = _data["cloud_area_fraction_low"];
            this.cloud_area_fraction_medium = _data["cloud_area_fraction_medium"];
            this.dew_point_temperature = _data["dew_point_temperature"];
            this.fog_area_fraction = _data["fog_area_fraction"];
            this.precipitation_amount = _data["precipitation_amount"];
            this.precipitation_amount_max = _data["precipitation_amount_max"];
            this.precipitation_amount_min = _data["precipitation_amount_min"];
            this.probability_of_precipitation = _data["probability_of_precipitation"];
            this.probability_of_thunder = _data["probability_of_thunder"];
            this.relative_humidity = _data["relative_humidity"];
            this.ultraviolet_index_clear_sky_max = _data["ultraviolet_index_clear_sky_max"];
            this.wind_from_direction = _data["wind_from_direction"];
            this.wind_speed = _data["wind_speed"];
            this.wind_speed_of_gust = _data["wind_speed_of_gust"];
        }
    }

    static fromJS(data: any): ForecastUnits {
        data = typeof data === 'object' ? data : {};
        let result = new ForecastUnits();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["air_pressure_at_sea_level"] = this.air_pressure_at_sea_level;
        data["air_temperature"] = this.air_temperature;
        data["air_temperature_max"] = this.air_temperature_max;
        data["air_temperature_min"] = this.air_temperature_min;
        data["cloud_area_fraction"] = this.cloud_area_fraction;
        data["cloud_area_fraction_high"] = this.cloud_area_fraction_high;
        data["cloud_area_fraction_low"] = this.cloud_area_fraction_low;
        data["cloud_area_fraction_medium"] = this.cloud_area_fraction_medium;
        data["dew_point_temperature"] = this.dew_point_temperature;
        data["fog_area_fraction"] = this.fog_area_fraction;
        data["precipitation_amount"] = this.precipitation_amount;
        data["precipitation_amount_max"] = this.precipitation_amount_max;
        data["precipitation_amount_min"] = this.precipitation_amount_min;
        data["probability_of_precipitation"] = this.probability_of_precipitation;
        data["probability_of_thunder"] = this.probability_of_thunder;
        data["relative_humidity"] = this.relative_humidity;
        data["ultraviolet_index_clear_sky_max"] = this.ultraviolet_index_clear_sky_max;
        data["wind_from_direction"] = this.wind_from_direction;
        data["wind_speed"] = this.wind_speed;
        data["wind_speed_of_gust"] = this.wind_speed_of_gust;
        return data;
    }
}

export interface IForecastUnits {
    air_pressure_at_sea_level?: string | undefined;
    air_temperature?: string | undefined;
    air_temperature_max?: string | undefined;
    air_temperature_min?: string | undefined;
    cloud_area_fraction?: string | undefined;
    cloud_area_fraction_high?: string | undefined;
    cloud_area_fraction_low?: string | undefined;
    cloud_area_fraction_medium?: string | undefined;
    dew_point_temperature?: string | undefined;
    fog_area_fraction?: string | undefined;
    precipitation_amount?: string | undefined;
    precipitation_amount_max?: string | undefined;
    precipitation_amount_min?: string | undefined;
    probability_of_precipitation?: string | undefined;
    probability_of_thunder?: string | undefined;
    relative_humidity?: string | undefined;
    ultraviolet_index_clear_sky_max?: string | undefined;
    wind_from_direction?: string | undefined;
    wind_speed?: string | undefined;
    wind_speed_of_gust?: string | undefined;
}
// ...
export class Data implements IData {
    /** Parameters which applies to this exact point in time */
    instant!: Instant;
    /** Parameters with validity times over twelve hours. Will not exist for all time steps. */
    next_12_hours?: Next_12_hours | undefined;
    /** Parameters with validity times over one hour. Will not exist for all time steps. */
    next_1_hours?: Next_1_hours | undefined;
    /** Parameters with validity times over six hours. Will not exist for all time steps. */
    next_6_hours?: Next_6_hours | undefined;

    constructor(data?: IData) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
        if (!data) {
            this.instant = new Instant();
        }
    }

    init(_data?: any) {
        if (_data) {
            this.instant = _data["instant"] ? Instant.fromJS(_data["instant"]) : new Instant();
            this.next_12_hours = _data["next_12_hours"] ? Next_12_hours.fromJS(_data["next_12_hours"]) : <any>undefined;
            this.next_1_hours = _data["next_1_hours"] ? Next_1_hours.fromJS(_data["next_1_hours"]) : <any>undefined;
            this.next_6_hours = _data["next_6_hours"] ? Next_6_hours.fromJS(_data["next_6_hours"]) : <any>undefined;
        }
    }

    static fromJS(data: any): Data {
        data = typeof data === 'object' ? data : {};
        let result = new Data();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["instant"] = this.instant ? this.instant.toJSON() : <any>undefined;
        data["next_12_hours"] = this.next_12_hours ? this.next_12_hours.toJSON() : <any>undefined;
        data["next_1_hours"] = this.next_1_hours ? this.next_1_hours.toJSON() : <any>undefined;
        data["next_6_hours"] = this.next_6_hours ? this.next_6_hours.toJSON() : <any>undefined;
        return data;
    }
}

export interface IData {
    /** Parameters which applies to this exact point in time */
    instant: Instant;
    /** Parameters with validity times over twelve hours. Will not exist for all time steps. */
    next_12_hours?: Next_12_hours | undefined;
    /** Parameters with validity times over one hour. Will not exist for all time steps. */
    next_1_hours?: Next_1_hours | undefined;
    /** Parameters with validity times over six hours. Will not exist for all time steps. */
    next_6_hours?: Next_6_hours | undefined;
}

export enum METJSONForecastType {
    Feature = "Feature",
}

export enum PointGeometryType {
    Point = "Point",
}

export class Instant implements IInstant {
    details?: ForecastTimeInstant | undefined;

    constructor(data?: IInstant) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.details = _data["details"] ? ForecastTimeInstant.fromJS(_data["details"]) : <any>undefined;
        }
    }

    static fromJS(data: any): Instant {
        data = typeof data === 'object' ? data : {};
        let result = new Instant();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["details"] = this.details ? this.details.toJSON() : <any>undefined;
        return data;
    }
}

export interface IInstant {
    details?: ForecastTimeInstant | undefined;
}

export class Next_12_hours implements INext_12_hours {
    details!: ForecastTimePeriod;
    summary!: ForecastSummary;

    constructor(data?: INext_12_hours) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
        if (!data) {
            this.details = new ForecastTimePeriod();
            this.summary = new ForecastSummary();
        }
    }

    init(_data?: any) {
        if (_data) {
            this.details = _data["details"] ? ForecastTimePeriod.fromJS(_data["details"]) : new ForecastTimePeriod();
            this.summary = _data["summary"] ? ForecastSummary.fromJS(_data["summary"]) : new ForecastSummary();
        }
    }

    static fromJS(data: any): Next_12_hours {
        data = typeof data === 'object' ? data : {};
        let result = new Next_12_hours();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["details"] = this.details ? this.details.toJSON() : <any>undefined;
        data["summary"] = this.summary ? this.summary.toJSON() : <any>undefined;
        return data;
    }
}